+++ changelog.md
@@ -1,4 +1,5 @@
-* Made symbol tabs and fonts data driven so they can be added, removed, customized with resourcepacks.
+* Made symbol tabs and fonts data driven, so they can be added, removed, customized with resourcepacks.
 For detailed documentation on the syntax, see [Resource pack syntax](https://github.com/replaceitem/symbol-chat/wiki/Resource-pack-syntax)
+* Added config option for changing the symbol button text colors
 * Added tooltip to the unicode table button
-* Some small UI adjustments and fixes 
+* Some small UI adjustments and fixes